requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pyodbc
azure-identity 
strawberry-graphql
starlette
#strawberry-graphql[debug-server]
#strawberry-graphql[asgi]
